Added standard bone length to reduce_bone_length_dispersion function plus minor changes
Added the option to use the standard bone length as the reference median in reduce_bone_length_dispersion function.
Created a dictionary with the bones lengths as proportions to the body height.
Added an input and variable to set the wanted body height in meters.
Deleted the custom_frame method for the add_rig function. Is not really helpful when you can use the median bone lenght.
Some minor changes to the custom mesh in add_mesh_to_rig function.